Problem: word "modules" are tightly coupled with the scheduler
Currently, they have to be manually integrated into the scheduler's pass and
every word has to be listed in it.

Solution: define a Module trait that allows to handle the entire
domain of words completely within one module. It also allows us
now to extend the schedule to allow for dynamic addition or
reconfiguraiton of modules
