+++ src/script/mod.rs
@@ -59,11 +59,6 @@
 //!   from carrying these references outside of the scope of the transaction)
 //!
 
-use num_bigint::{BigUint, BigInt, Sign, };
-use num_traits::{Zero, One, Signed};
-use num_traits::ToPrimitive;
-use core::ops::{Add, Sub};
-
 use std::collections::BTreeMap;
 
 pub mod envheap;
@@ -85,98 +80,11 @@ macro_rules! word {
     )
 }
 
-// Built-in words
-// TODO: the list of built-in words is far from completion
-
-// How to write a new built-in word:
-// 1. Add `word!(...)` to define a constant
-// 2. Document it
-// 3. Write a test in mod tests
-// 4. Add `handle_word` function in VM and list it in `match_words!()` macro
-//    invocation in VM::pass
-
-// Category: Stack
-word!(DROP, (a => ), b"\x84DROP");
-word!(DUP, (a => a, a), b"\x83DUP");
-word!(SWAP, (a, b => b, a), b"\x84SWAP");
-word!(TWOSWAP, (a, b, c, d => c, d, a, b), b"\x852SWAP");
-word!(ROT, (a, b, c  => b, c, a), b"\x83ROT");
-word!(TWOROT, (a, b, c, d, e, f  => c, d, e, f, a, b), b"\x842ROT");
-word!(OVER, (a, b => a, b, a), b"\x84OVER");
-word!(TWOOVER, (a, b, c, d => a, b, c, d, a, b), b"\x852OVER");
-word!(DEPTH, b"\x85DEPTH");
-word!(UNWRAP, b"\x86UNWRAP");
-word!(WRAP, b"\x84WRAP");
-
-// Category: Byte arrays
-word!(EQUALQ, (a, b => c), b"\x86EQUAL?");
-word!(LTQ, (a, b => c), b"\x83LT?");
-word!(GTQ, (a, b => c), b"\x83GT?");
-word!(LENGTH, (a => b), b"\x86LENGTH");
-word!(CONCAT, (a, b => c), b"\x86CONCAT");
-word!(SLICE, (a, b, c => d), b"\x85SLICE");
-word!(PAD, (a, b, c => d), b"\x83PAD");
-
-// Category: arithmetics
-word!(UINT_ADD, (a, b => c), b"\x88UINT/ADD");
-word!(UINT_SUB, (a, b => c), b"\x88UINT/SUB");
-word!(INT_ADD, (a, b => c), b"\x87INT/ADD");
-word!(INT_SUB, (a, b => c), b"\x87INT/SUB");
-
-// Casting
-word!(INT_TO_UINT, (a => b), b"\x89INT->UINT");
-word!(UINT_TO_INT, (a => b), b"\x89UINT->INT");
-
-// Category: Control flow
-#[cfg(feature = "scoped_dictionary")]
-word!(EVAL_SCOPED, b"\x8BEVAL/SCOPED");
-#[cfg(feature = "scoped_dictionary")]
-word!(SCOPE_END, b"\x80\x8BEVAL/SCOPED"); // internal word
-word!(DOWHILE, b"\x87DOWHILE");
-word!(TIMES, b"\x85TIMES");
-word!(EVAL, b"\x84EVAL");
-word!(EVAL_VALIDP, b"\x8BEVAL/VALID?");
 word!(TRY, b"\x83TRY");
 word!(TRY_END, b"\x80\x83TRY"); // internal word
-word!(SET, b"\x83SET");
-word!(DEF, b"\x83DEF");
-word!(IF, b"\x82IF"); // for reference, implemented in builtins
-word!(IFELSE, b"\x86IFELSE");
 
-// Category: Logical operations
-word!(NOT, (a => c), b"\x83NOT");
-word!(AND, (a, b => c), b"\x83AND");
-word!(OR, (a, b => c), b"\x82OR");
-
-// Category: pubsub
-word!(SEND, (a => ), b"\x84SEND");
-
-// Category: experimental features
-word!(FEATUREQ, (a => b), b"\x88FEATURE?");
 
 use std::str;
-
-// Builtin words that are implemented in PumpkinScript
-lazy_static! {
-  static ref BUILTIN_FILE: &'static [u8] = include_bytes!("builtins");
-
-  static ref BUILTIN_DEFS: Vec<Vec<u8>> = textparser::programs(*BUILTIN_FILE).unwrap().1;
-
-  static ref BUILTINS: BTreeMap<&'static [u8], Vec<u8>> = {
-      let mut map = BTreeMap::new();
-      let ref defs : Vec<Vec<u8>> = *BUILTIN_DEFS;
-      for definition in defs {
-          match binparser::word(definition.as_slice()) {
-              nom::IResult::Done(&[0x81, b':', ref rest..], _) => {
-                  let word = &definition[0..definition.len() - rest.len() - 2];
-                  map.insert(word, Vec::from(rest));
-              },
-              other => panic!("builtin definition parse error {:?}", other)
-          }
-      }
-      map
-  };
-}
 
 // To add words that don't belong to a core set,
 // add a module with a handler, and reference it in the VM's pass
@@ -383,6 +291,7 @@ impl<'a> Env<'a> {
             self.dictionary.push(BTreeMap::new());
         }
     }
+
 }
 
 
@@ -397,19 +306,6 @@ pub fn offset_by_size(size: usize) -> usize {
         65536...4294967296 => 5,
         _ => unreachable!(),
     }
-}
-
-pub fn bytes_to_bigint(bytes: &[u8]) -> Option<BigInt> {
-    if bytes.len() >= 2 {
-        match bytes[0] {
-            0x00 => Some(Sign::Minus),
-            0x01 => Some(Sign::Plus),
-            _ => None
-        }.and_then(|sign| Some(BigInt::from_bytes_be(sign, &bytes[1..])))
-    } else {
-        None
-    }
-
 }
 
 include!("macros.rs");
@@ -455,6 +351,12 @@ pub mod timestamp_hlc;
 pub mod hash;
 pub mod json;
 
+pub trait Module<'a> {
+    fn init(&mut self, _: &mut Env<'a>, _: EnvId) {}
+    fn done(&mut self, _: &mut Env<'a>, _: EnvId) {}
+    fn handle(&mut self, env: &mut Env<'a>, word: &'a [u8], pid: EnvId) -> PassResult<'a>;
+}
+
 /// VM is a PumpkinScript scheduler and interpreter. This is the
 /// most central part of this module.
 ///
@@ -492,11 +394,7 @@ use std::collections::VecDeque;
 pub struct VM<'a> {
     inbox: Receiver<RequestMessage>,
     sender: Sender<RequestMessage>,
-    publisher: pubsub::PublisherAccessor<Vec<u8>>,
-    storage: storage::Handler<'a>,
-    hlc: timestamp_hlc::Handler<'a>,
-    hash: hash::Handler<'a>,
-    json: json::Handler<'a>,
+    modules: Vec<Box<Module<'a> + 'a>>,
 }
 
 unsafe impl<'a> Send for VM<'a> {}
@@ -527,11 +425,12 @@ impl<'a> VM<'a> {
         VM {
             inbox: receiver,
             sender: sender.clone(),
-            publisher: publisher,
-            storage: storage::Handler::new(db_env, db),
-            hlc: timestamp_hlc::Handler::new(),
-            hash: hash::Handler::new(),
-            json: json::Handler::new(),
+            modules: vec![Box::new(basic::Handler::new(publisher)),
+                          Box::new(storage::Handler::new(db_env, db)),
+                          Box::new(hash::Handler::new()),
+                          Box::new(timestamp_hlc::Handler::new()),
+                          Box::new(json::Handler::new()),
+            ],
         }
     }
 
@@ -552,14 +451,18 @@ impl<'a> VM<'a> {
         let mut envs: VecDeque<(EnvId, Env<'a>, Sender<ResponseMessage>)> = VecDeque::new();
 
         loop {
-
             match envs.pop_front() {
                 Some((pid, mut env, chan)) => {
+                    let program = env.program[env.program.len() - 1];
                     match self.pass(&mut env, pid.clone()) {
                         Err(Error::Reschedule) => {
+                            env.program.push(program);
                             envs.push_back((pid, env, chan));
                         },
                         Err(err) => {
+                            for mut module in self.modules.iter_mut() {
+                                module.done(&mut env, pid);
+                            }
                             let _ = chan.send(ResponseMessage::EnvFailed(pid,
                                                                          err,
                                                                          Some(env.stack_copy()),
@@ -567,6 +470,9 @@ impl<'a> VM<'a> {
                         }
                         Ok(()) => {
                             if env.program.is_empty() || (env.program.len() == 1 && env.program[0].len() == 0) {
+                                for mut module in self.modules.iter_mut() {
+                                    module.done(&mut env, pid);
+                                }
                                 let _ = chan.send(ResponseMessage::EnvTerminated(pid,
                                                                                  env.stack_copy(),
                                                                                  env.stack_size));
@@ -585,7 +491,7 @@ impl<'a> VM<'a> {
                 if let Err(mpsc::TryRecvError::Empty) = msg {
                     continue;
                 }
-                msg.map_err(|_| mpsc::RecvError{})
+                msg.map_err(|_| mpsc::RecvError {})
             };
             match message {
                 Err(err) => panic!("error receiving: {:?}", err),
@@ -597,6 +503,9 @@ impl<'a> VM<'a> {
                                 Ok(slice) => {
                                     slice.copy_from_slice(program.as_slice());
                                     env.program.push(slice);
+                                    for mut module in self.modules.iter_mut() {
+                                        module.init(&mut env, pid);
+                                    }
                                     envs.push_back((pid, env, chan));
                                 }
                                 Err(err) => {
@@ -655,830 +564,40 @@ impl<'a> VM<'a> {
             if rest.len() > 0 {
                 env.program.push(rest);
             }
-            handle_words!(self, env,
-                          program,
-                          word,
-                          new_env,
-                          pid,
-                          {self => handle_builtins,
-                           self => handle_drop,
-                           self => handle_dup,
-                           self => handle_swap,
-                           self => handle_2swap,
-                           self => handle_rot,
-                           self => handle_2rot,
-                           self => handle_over,
-                           self => handle_2over,
-                           self => handle_depth,
-                           self => handle_wrap,
-                           self => handle_ltp,
-                           self => handle_gtp,
-                           self => handle_equal,
-                           self => handle_concat,
-                           self => handle_slice,
-                           self => handle_pad,
-                           self => handle_uint_add,
-                           self => handle_uint_sub,
-                           self => handle_int_add,
-                           self => handle_int_sub,
-                           self => handle_int_to_uint,
-                           self => handle_uint_to_int,
-                           self => handle_length,
-                           self => handle_dowhile,
-                           self => handle_times,
-                           self => handle_scope_end,
-                           self => handle_eval,
-                           self => handle_eval_validp,
-                           self => handle_eval_scoped,
-                           self => handle_try,
-                           self => handle_try_end,
-                           self => handle_unwrap,
-                           self => handle_set,
-                           self => handle_def,
-                           self => handle_not,
-                           self => handle_and,
-                           self => handle_or,
-                           self => handle_ifelse,
-                           // storage
-                           self.storage => handle_write,
-                           self.storage => handle_read,
-                           self.storage => handle_assoc,
-                           self.storage => handle_assocq,
-                           self.storage => handle_retr,
-                           self.storage => handle_commit,
-                           self.storage => handle_cursor,
-                           self.storage => handle_cursor_first,
-                           self.storage => handle_cursor_next,
-                           self.storage => handle_cursor_prev,
-                           self.storage => handle_cursor_last,
-                           self.storage => handle_cursor_seek,
-                           self.storage => handle_cursor_cur,
-                           // timestamping
-                           self.hlc => handle_hlc,
-                           self.hlc => handle_hlc_lc,
-                           self.hlc => handle_hlc_tick,
-                           // hashing
-                           self.hash => handle_hash_sha1,
-                           self.hash => handle_hash_sha224,
-                           self.hash => handle_hash_sha256,
-                           self.hash => handle_hash_sha384,
-                           self.hash => handle_hash_sha512,
-                           self.hash => handle_hash_sha512_224,
-                           self.hash => handle_hash_sha512_256,
-                           // json
-                           self.json => handle_jsonq,
-                           self.json => handle_json_objectq,
-                           self.json => handle_json_stringq,
-                           self.json => handle_json_numberq,
-                           self.json => handle_json_booleanq,
-                           self.json => handle_json_arrayq,
-                           self.json => handle_json_nullq,
-                           self.json => handle_json_get,
-                           self.json => handle_json_hasq,
-                           self.json => handle_json_set,
-                           self.json => handle_json_string_to,
-                           self.json => handle_json_to_string,
-                           // pubsub
-                           self => handle_send,
-                           // features
-                           self => handle_featurep,
-                           // catch-all (NB: keep it last)
-                           self => handle_dictionary
-                           })
+            if word != TRY_END && !env.aborting_try.is_empty() {
+                return Ok(())
+            }
+
+            try_word!(env, self.handle_try(env, word, pid));
+            try_word!(env, self.handle_try_end(env, word, pid));
+
+            for mut module in self.modules.iter_mut() {
+                try_word!(env, module.handle(env, word, pid));
+            }
+
+            // catch-all (NB: keep it last)
+            try_word!(env, self.handle_dictionary(env, word, pid));
+
+            // if nothing worked...
+            handle_error!(env, error_unknown_word!(word))
         } else {
             handle_error!(env, error_decoding!())
         }
     }
 
+
     #[inline]
-    fn handle_builtins(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        if BUILTINS.contains_key(word) {
-            let vec = BUILTINS.get(word).unwrap();
-            env.program.push(vec.as_slice());
+    #[cfg(not(feature = "scoped_dictionary"))]
+    fn handle_dictionary(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
+        if env.dictionary.contains_key(word) {
+            {
+                let def = env.dictionary.get(word).unwrap();
+                env.program.push(def);
+            }
             Ok(())
         } else {
             Err(Error::UnknownWord)
         }
-    }
-
-    #[inline]
-    fn handle_dup(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, DUP);
-        let v = stack_pop!(env);
-
-        env.push(v);
-        env.push(v);
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_swap(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, SWAP);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-
-        env.push(a);
-        env.push(b);
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_2swap(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, TWOSWAP);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-        let c = stack_pop!(env);
-        let d = stack_pop!(env);
-
-        env.push(b);
-        env.push(a);
-
-        env.push(d);
-        env.push(c);
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_over(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, OVER);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-
-        env.push(b);
-        env.push(a);
-        env.push(b);
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_2over(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, TWOOVER);
-        let d = stack_pop!(env);
-        let c = stack_pop!(env);
-        let b = stack_pop!(env);
-        let a = stack_pop!(env);
-
-        env.push(a);
-        env.push(b);
-        env.push(c);
-        env.push(d);
-        env.push(a);
-        env.push(b);
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_rot(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, ROT);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-        let c = stack_pop!(env);
-
-        env.push(b);
-        env.push(a);
-        env.push(c);
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_2rot(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, TWOROT);
-        let f = stack_pop!(env);
-        let e = stack_pop!(env);
-        let d = stack_pop!(env);
-        let c = stack_pop!(env);
-        let b = stack_pop!(env);
-        let a = stack_pop!(env);
-
-        env.push(c);
-        env.push(d);
-        env.push(e);
-        env.push(f);
-        env.push(a);
-        env.push(b);
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_drop(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, DROP);
-        let _ = stack_pop!(env);
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_depth(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, DEPTH);
-        let bytes = BigUint::from(env.stack_size).to_bytes_be();
-        let slice = alloc_and_write!(bytes.as_slice(), env);
-        env.push(slice);
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_wrap(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, WRAP);
-        let n = stack_pop!(env);
-
-        let mut n_int = BigUint::from_bytes_be(n).to_u64().unwrap() as usize;
-
-        let mut vec = Vec::new();
-
-        while n_int > 0 {
-            let item = stack_pop!(env);
-            vec.insert(0, item);
-            n_int -= 1;
-        }
-
-        let size = vec.clone().into_iter()
-            .fold(0, |a, item| a + item.len() + offset_by_size(item.len()));
-
-        let mut slice = alloc_slice!(size, env);
-
-        let mut offset = 0;
-        for item in vec {
-            write_size_into_slice!(item.len(), &mut slice[offset..]);
-            offset += offset_by_size(item.len());
-            slice[offset..offset + item.len()].copy_from_slice(item);
-            offset += item.len();
-        }
-        env.push(slice);
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_equal(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, EQUALQ);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-
-        if a == b {
-            env.push(STACK_TRUE);
-        } else {
-            env.push(STACK_FALSE);
-        }
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_not(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, NOT);
-        let a = stack_pop!(env);
-
-        if a == STACK_TRUE {
-            env.push(STACK_FALSE);
-        } else if a == STACK_FALSE {
-            env.push(STACK_TRUE);
-        } else {
-            return Err(error_invalid_value!(a));
-        }
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_and(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, AND);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-
-        if !(a == STACK_TRUE || a == STACK_FALSE) {
-            return Err(error_invalid_value!(a));
-        }
-        if !(b == STACK_TRUE || b == STACK_FALSE) {
-            return Err(error_invalid_value!(b));
-        }
-
-        if a == STACK_TRUE && b == STACK_TRUE {
-            env.push(STACK_TRUE);
-        } else if a == STACK_FALSE || b == STACK_FALSE {
-            env.push(STACK_FALSE);
-        }
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_or(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, OR);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-
-        if !(a == STACK_TRUE || a == STACK_FALSE) {
-            return Err(error_invalid_value!(a));
-        }
-        if !(b == STACK_TRUE || b == STACK_FALSE) {
-            return Err(error_invalid_value!(b));
-        }
-
-        if a == STACK_TRUE || b == STACK_TRUE {
-            env.push(STACK_TRUE);
-        } else {
-            env.push(STACK_FALSE);
-        }
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_ifelse(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, IFELSE);
-        let else_ = stack_pop!(env);
-        let then = stack_pop!(env);
-        let cond = stack_pop!(env);
-
-        if cond == STACK_TRUE {
-            env.program.push(then);
-            Ok(())
-        } else if cond == STACK_FALSE {
-            env.program.push(else_);
-            Ok(())
-        } else {
-            Err(error_invalid_value!(cond))
-        }
-    }
-
-    #[inline]
-    fn handle_ltp(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, LTQ);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-
-        if b < a {
-            env.push(STACK_TRUE);
-        } else {
-            env.push(STACK_FALSE);
-        }
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_gtp(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, GTQ);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-
-        if b > a {
-            env.push(STACK_TRUE);
-        } else {
-            env.push(STACK_FALSE);
-        }
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_concat(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, CONCAT);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-
-        let slice = alloc_slice!(a.len() + b.len(), env);
-
-        slice[0..b.len()].copy_from_slice(b);
-        slice[b.len()..b.len()+a.len()].copy_from_slice(a);
-
-        env.push(slice);
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_slice(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, SLICE);
-        let end = stack_pop!(env);
-        let start = stack_pop!(env);
-        let slice = stack_pop!(env);
-
-        let start_int = BigUint::from_bytes_be(start).to_u64().unwrap() as usize;
-        let end_int = BigUint::from_bytes_be(end).to_u64().unwrap() as usize;
-
-        // range conditions
-        if start_int > end_int {
-            return Err(error_invalid_value!(start));
-        }
-
-        if start_int > slice.len() - 1 {
-            return Err(error_invalid_value!(start));
-        }
-
-        if end_int > slice.len() {
-            return Err(error_invalid_value!(end));
-        }
-
-        env.push(&slice[start_int..end_int]);
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_pad(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, PAD);
-        let byte = stack_pop!(env);
-        let size = stack_pop!(env);
-        let value = stack_pop!(env);
-
-        if byte.len() != 1 {
-            return Err(error_invalid_value!(byte));
-        }
-
-        let size_int = BigUint::from_bytes_be(size).to_u64().unwrap() as usize;
-
-        if size_int > 1024 {
-            return Err(error_invalid_value!(size));
-        }
-
-        let slice = alloc_slice!(size_int, env);
-
-        for i in 0..size_int-value.len() {
-            slice[i] = byte[0];
-        }
-        slice[size_int-value.len()..].copy_from_slice(value);
-
-        env.push(slice);
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_length(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, LENGTH);
-        let a = stack_pop!(env);
-
-        let len = BigUint::from(a.len() as u64);
-        let len_bytes = len.to_bytes_be();
-
-        let slice = alloc_and_write!(len_bytes.as_slice(), env);
-
-        env.push(slice);
-
-        Ok(())
-    }
-
-    #[inline]
-    #[cfg(feature = "scoped_dictionary")]
-    fn handle_eval_scoped(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, EVAL_SCOPED);
-        env.push_dictionary();
-        let a = stack_pop!(env);
-        env.program.push(SCOPE_END);
-        env.program.push(a);
-        Ok(())
-    }
-
-    #[inline]
-    #[cfg(not(feature = "scoped_dictionary"))]
-    fn handle_eval_scoped(&mut self, _: &Env<'a>, _: &'a [u8], _: EnvId) -> PassResult<'a> {
-        Err(Error::UnknownWord)
-    }
-
-
-    #[inline]
-    #[cfg(feature = "scoped_dictionary")]
-    fn handle_scope_end(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, SCOPE_END);
-        env.pop_dictionary();
-        Ok(())
-    }
-
-
-    #[inline]
-    #[cfg(not(feature = "scoped_dictionary"))]
-    fn handle_scope_end(&mut self, _: &mut Env<'a>, _: &'a [u8], _: EnvId) -> PassResult<'a> {
-        Err(Error::UnknownWord)
-    }
-
-    #[inline]
-    fn handle_uint_add(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, UINT_ADD);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-
-        let a_uint = BigUint::from_bytes_be(a);
-        let b_uint = BigUint::from_bytes_be(b);
-
-        let c_uint = a_uint.add(b_uint);
-
-        let c_bytes = c_uint.to_bytes_be();
-
-        let slice = alloc_and_write!(c_bytes.as_slice(), env);
-        env.push(slice);
-        Ok(())
-    }
-
-    fn handle_int_add(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, INT_ADD);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-
-        let a_int = bytes_to_bigint(a);
-        let b_int = bytes_to_bigint(b);
-
-        if a_int == None {
-            return Err(error_invalid_value!(a))
-        }
-        if b_int == None {
-            return Err(error_invalid_value!(b))
-        }
-
-        let c_int = a_int.unwrap().add(b_int.unwrap());
-
-        let mut bytes = if c_int.is_negative() {
-            vec![0x00]
-        } else {
-            vec![0x01]
-        };
-        let (_, c_bytes) = c_int.to_bytes_be();
-        bytes.extend_from_slice(&c_bytes);
-        let slice = alloc_and_write!(bytes.as_slice(), env);
-        env.push(slice);
-        Ok(())
-    }
-
-    fn handle_int_sub(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, INT_SUB);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-
-        let a_int = bytes_to_bigint(a);
-        let b_int = bytes_to_bigint(b);
-
-        if a_int == None {
-            return Err(error_invalid_value!(a))
-        }
-        if b_int == None {
-            return Err(error_invalid_value!(b))
-        }
-
-        let c_int = b_int.unwrap().sub(a_int.unwrap());
-
-        let mut bytes = if c_int.is_negative() {
-            vec![0x00]
-        } else {
-            vec![0x01]
-        };
-        let (_, c_bytes) = c_int.to_bytes_be();
-        bytes.extend_from_slice(&c_bytes);
-        let slice = alloc_and_write!(bytes.as_slice(), env);
-        env.push(slice);
-        Ok(())
-    }
-
-    fn handle_int_to_uint(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId)
-        -> PassResult<'a> {
-        word_is!(env, word, INT_TO_UINT);
-        let a = stack_pop!(env);
-        let a_int = bytes_to_bigint(a);
-
-        if a_int == None {
-            return Err(error_invalid_value!(a))
-        }
-
-        match a_int.unwrap().to_biguint() {
-            Some(a_uint) => {
-                let a_bytes = a_uint.to_bytes_be();
-                let slice = alloc_and_write!(a_bytes.as_slice(), env);
-                env.push(slice);
-                Ok(())
-            },
-            None => {
-                Err(error_invalid_value!(a))
-            }
-        }
-    }
-
-    fn handle_uint_to_int(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId)
-                          -> PassResult<'a> {
-        word_is!(env, word, UINT_TO_INT);
-        let a = stack_pop!(env);
-        let a_uint = BigUint::from_bytes_be(a);
-
-        let mut bytes = vec![0x01];
-        let a_bytes = a_uint.to_bytes_be();
-        bytes.extend_from_slice(&a_bytes);
-        let slice = alloc_and_write!(bytes.as_slice(), env);
-
-        env.push(slice);
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_uint_sub(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, UINT_SUB);
-        let a = stack_pop!(env);
-        let b = stack_pop!(env);
-
-        let a_uint = BigUint::from_bytes_be(a);
-        let b_uint = BigUint::from_bytes_be(b);
-
-        if a_uint > b_uint {
-            return Err(error_invalid_value!(a));
-        }
-
-        let c_uint = b_uint.sub(a_uint);
-
-        let c_bytes = c_uint.to_bytes_be();
-        let slice = alloc_and_write!(c_bytes.as_slice(), env);
-        env.push(slice);
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_eval(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, EVAL);
-        let a = stack_pop!(env);
-        env.program.push(a);
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_eval_validp(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, EVAL_VALIDP);
-        let a = stack_pop!(env);
-        if parse_bin(a).is_ok() {
-            env.push(STACK_TRUE);
-        } else {
-            env.push(STACK_FALSE);
-        }
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_try(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, TRY);
-        let v = stack_pop!(env);
-        env.tracking_errors += 1;
-        env.program.push(TRY_END);
-        env.program.push(v);
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_try_end(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, TRY_END);
-        env.tracking_errors -= 1;
-        if env.aborting_try.is_empty() {
-            env.push(_EMPTY);
-            Ok(())
-        } else if let Some(Error::ProgramError(err)) = env.aborting_try.pop() {
-            let slice = alloc_and_write!(err.as_slice(), env);
-            env.push(slice);
-            Ok(())
-        } else {
-            env.push(_EMPTY);
-            Ok(())
-        }
-    }
-
-    #[inline]
-    fn handle_unwrap(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, UNWRAP);
-        let mut current = stack_pop!(env);
-        while current.len() > 0 {
-            match binparser::data(current) {
-                nom::IResult::Done(rest, val) => {
-                    env.push(&val[offset_by_size(val.len())..]);
-                    current = rest
-                },
-                _ => {
-                    return Err(error_invalid_value!(current))
-                }
-            }
-        }
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_dowhile(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, DOWHILE);
-        let v = stack_pop!(env);
-
-        let mut vec = Vec::new();
-
-        let mut header = vec![0;offset_by_size(v.len() + DOWHILE.len() + offset_by_size(v.len()))];
-        write_size_into_slice!(offset_by_size(v.len()) + v.len() + DOWHILE.len(), header.as_mut_slice());
-        vec.append(&mut header);
-
-        // inject code closure size
-        let mut header = vec![0;offset_by_size(v.len())];
-        write_size_into_slice!(v.len(), header.as_mut_slice());
-        vec.append(&mut header);
-
-        // inject code closure
-        vec.extend_from_slice(v);
-        // inject DOWHILE
-        vec.extend_from_slice(DOWHILE);
-        // inject IF
-        vec.extend_from_slice(IF);
-
-        let slice = alloc_and_write!(vec.as_slice(), env);
-        env.program.push(slice);
-        env.program.push(v);
-
-        Ok(())
-    }
-
-    #[inline]
-    fn handle_times(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, TIMES);
-        let count = stack_pop!(env);
-
-        let v = stack_pop!(env);
-
-        let counter = BigUint::from_bytes_be(count);
-        if counter.is_zero() {
-            Ok(())
-        } else {
-            let mut vec = Vec::new();
-            if counter != BigUint::one() {
-                // inject the prefix for the code
-                let mut header = vec![0;offset_by_size(v.len())];
-                write_size_into_slice!(v.len(), header.as_mut_slice());
-                vec.append(&mut header);
-                vec.extend_from_slice(v);
-                // inject the decremented counter
-                let counter = counter.sub(BigUint::one());
-                let mut counter_bytes = counter.to_bytes_be();
-                let mut header = vec![0;offset_by_size(counter_bytes.len())];
-                write_size_into_slice!(counter_bytes.len(), header.as_mut_slice());
-                vec.append(&mut header);
-                vec.append(&mut counter_bytes);
-                // inject TIMES
-                vec.extend_from_slice(TIMES);
-            }
-            let slice = alloc_and_write!(vec.as_slice(), env);
-            env.program.push(slice);
-            env.program.push(v);
-            Ok(())
-        }
-    }
-
-    #[inline]
-    fn handle_set(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, SET);
-        let word = stack_pop!(env);
-        let value = stack_pop!(env);
-        match binparser::word(word) {
-            nom::IResult::Done(_, _) => {
-                let slice = alloc_slice!(value.len() + offset_by_size(value.len()), env);
-                write_size_into_slice!(value.len(), slice);
-                let offset = offset_by_size(value.len());
-                slice[offset..offset + value.len()].copy_from_slice(value);
-                #[cfg(feature = "scoped_dictionary")]
-                {
-                    let mut dict = env.dictionary.pop().unwrap();
-                    dict.insert(word, slice);
-                    env.dictionary.push(dict);
-                }
-                #[cfg(not(feature = "scoped_dictionary"))]
-                env.dictionary.insert(word, slice);
-                Ok(())
-            },
-            _ => Err(error_invalid_value!(word))
-        }
-    }
-
-    fn handle_def(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, DEF);
-        let word = stack_pop!(env);
-        let value = stack_pop!(env);
-        match binparser::word(word) {
-            nom::IResult::Done(_, _) => {
-                #[cfg(feature = "scoped_dictionary")]
-                {
-                    let mut dict = env.dictionary.pop().unwrap();
-                    dict.insert(word, value);
-                    env.dictionary.push(dict);
-                }
-                #[cfg(not(feature = "scoped_dictionary"))]
-                env.dictionary.insert(word, value);
-                Ok(())
-            },
-            _ => Err(error_invalid_value!(word))
-        }
-    }
-
-
-    #[inline]
-    fn handle_send(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, SEND);
-        let topic = stack_pop!(env);
-        let data = stack_pop!(env);
-
-        let receiver = self.publisher.send_async(Vec::from(topic), Vec::from(data));
-
-        env.send_ack = Some(receiver);
-
-        Ok(())
     }
 
     #[inline]
@@ -1499,40 +618,37 @@ impl<'a> VM<'a> {
     }
 
     #[inline]
-    #[cfg(not(feature = "scoped_dictionary"))]
-    fn handle_dictionary(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        if env.dictionary.contains_key(word) {
-            {
-                let def = env.dictionary.get(word).unwrap();
-                env.program.push(def);
-            }
-            Ok(())
-        } else {
-            Err(Error::UnknownWord)
-        }
-    }
-
-    #[inline]
-    #[allow(unused_variables)]
-    fn handle_featurep(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
-        word_is!(env, word, FEATUREQ);
-        let name = stack_pop!(env);
-
-        #[cfg(feature = "scoped_dictionary")]
-        {
-            if name == "scoped_dictionary".as_bytes() {
-                env.push(STACK_TRUE);
-                return Ok(())
-            }
-        }
-
-        env.push(STACK_FALSE);
-
+    fn handle_try(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
+        word_is!(env, word, TRY);
+        let v = stack_pop!(env);
+        env.tracking_errors += 1;
+        env.program.push(TRY_END);
+        env.program.push(v);
         Ok(())
     }
 
+    #[inline]
+    fn handle_try_end(&mut self, env: &mut Env<'a>, word: &'a [u8], pid: EnvId) -> PassResult<'a> {
+        word_is!(env, word, TRY_END);
+        env.tracking_errors -= 1;
+        if env.aborting_try.is_empty() {
+            env.push(_EMPTY);
+            Ok(())
+        } else if let Some(Error::ProgramError(err)) = env.aborting_try.pop() {
+            for mut module in self.modules.iter_mut() {
+                module.done(env, pid);
+            }
+            let slice = alloc_and_write!(err.as_slice(), env);
+            env.push(slice);
+            Ok(())
+        } else {
+            env.push(_EMPTY);
+            Ok(())
+        }
+    }
 }
 
+pub mod basic;
 pub mod compose;
 
 #[cfg(test)]
@@ -1542,7 +658,6 @@ mod tests {
     use script::{Env, VM, Error, RequestMessage, ResponseMessage, EnvId, parse, offset_by_size};
     use std::sync::mpsc;
     use std::fs;
-    use std::thread;
     use tempdir::TempDir;
     use lmdb;
     use crossbeam;
@@ -1570,49 +685,6 @@ mod tests {
         assert!(env.stack.len() >= target);
     }
 
-    use std::time::Duration;
-
-    #[test]
-    fn send() {
-        eval!("\"Hello\" \"Topic\" SEND", env, result, publisher_accessor, {
-            let (sender1, receiver1) = mpsc::channel();
-            publisher_accessor.subscribe(Vec::from("Topic"), sender1);
-
-            let (sender0, receiver0) = mpsc::channel();
-            thread::spawn(move ||  {
-               match receiver1.recv() {
-                  Ok((topic, message, callback)) => {
-                     callback.send(());
-                     sender0.send((topic, message));
-                  },
-                  e => panic!("unexpected result {:?}", e)
-               };
-
-            });
-
-        }, {
-            assert!(!result.is_err());
-
-            let result = receiver0.recv_timeout(Duration::from_secs(1)).unwrap();
-            assert_eq!(result, (Vec::from("Topic"), Vec::from("Hello")));
-        });
-
-        eval!("\"Hello\" \"Topic1\" SEND", env, result, publisher_accessor, {
-            let (sender, receiver) = mpsc::channel();
-            publisher_accessor.subscribe(Vec::from("Topic"), sender);
-        }, {
-            assert!(!result.is_err());
-            assert!(receiver.recv_timeout(Duration::from_secs(1)).is_err());
-        });
-
-        eval!("\"Topic\" SEND", env, result, {
-            assert_error!(result, "[\"Empty stack\" [] 4]");
-        });
-
-        eval!("SEND", env, result, {
-            assert_error!(result, "[\"Empty stack\" [] 4]");
-        });
-    }
 
     #[test]
     fn unknown_word() {
@@ -1628,54 +700,7 @@ mod tests {
         });
     }
 
-    #[test]
-    fn try() {
-        eval!("[1 DUP] TRY", env, result, {
-            assert_eq!(Vec::from(env.pop().unwrap()), parsed_data!("[]"));
-            assert_eq!(Vec::from(env.pop().unwrap()), parsed_data!("0x01"));
-            assert_eq!(Vec::from(env.pop().unwrap()), parsed_data!("0x01"));
-            assert_eq!(env.pop(), None);
-        });
-
-        eval!("[DUP] TRY", env, result, {
-            assert!(!result.is_err());
-            assert_eq!(Vec::from(env.pop().unwrap()), parsed_data!("[\"Empty stack\" [] 4]"));
-            assert_eq!(env.pop(), None);
-        });
-
-        eval!("[NOTAWORD] TRY", env, result, {
-            assert_eq!(Vec::from(env.pop().unwrap()), parsed_data!("[\"Unknown word: NOTAWORD\" ['NOTAWORD] 2]"));
-            assert_eq!(env.pop(), None);
-        });
-
-        eval!("[[DUP] TRY 0x20 NOT] TRY", env, result, {
-            assert!(!result.is_err());
-            assert_eq!(Vec::from(env.pop().unwrap()), parsed_data!("[\"Invalid value\" [0x20] 3]"));
-            assert_eq!(Vec::from(env.pop().unwrap()), parsed_data!("[\"Empty stack\" [] 4]"));
-            assert_eq!(env.pop(), None);
-        });
-
-        eval!("[1 DUP] TRY STACK DROP DUP", env, result, {
-            assert!(result.is_err());
-        });
-
-        eval!("[DUP] TRY STACK DROP DUP", env, result, {
-            assert!(result.is_err());
-        });
-
-        eval!("1 TRY", env, result, {
-            assert_eq!(Vec::from(env.pop().unwrap()), parsed_data!("[\"Decoding error\" [] 5]"));
-            assert_eq!(env.pop(), None);
-        });
-
-    }
-
     use test::Bencher;
-
-    #[bench]
-    fn times(b: &mut Bencher) {
-       bench_eval!("[1 DROP] 1000 TIMES", b);
-    }
 
     #[bench]
     fn ackermann(b: &mut Bencher) { // HT @5HT
@@ -1684,7 +709,5 @@ mod tests {
         'ack DEF \
         3 4 ack", b);
     }
-
-
 
 }
